requirement.txt
appdirs==1.4.4
ApiRetrys==0.0.3
beautifulsoup4==4.12.3
asttokens==2.4.1
botocore==1.34.93
boto3==1.34.93
certifi==2024.2.2
bs4==0.0.2
click==8.1.7
charset-normalizer==3.3.2
cssselect==1.2.0
colorama==0.4.6
executing==2.0.1
dekimashita==0.0.3
filelock==3.13.4
fake-useragent==1.5.1
greenlet==3.0.3
gdown==5.1.0
idna==3.7
icecream==2.1.3
jmespath==1.0.1
importlib_metadata==7.1.0
kafka-python-ng==2.2.2
kafka-python==2.0.2
lxml==5.2.1
logging==0.4.9.6
numpy==1.26.4
lxml_html_clean==0.1.1
parse==1.20.1
pandas==2.2.2
pip==24.0
patool==2.2.0
pyee==11.1.0
playwright==1.43.0
pyppeteer==2.0.0
Pygments==2.17.2
PySocks==1.7.1
pyquery==2.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
rarfile==4.2
requests==2.31.0
requests-html==0.10.0
s3transfer==0.10.1
six==1.16.0
soupsieve==2.5
tqdm==4.66.2
typing_extensions==4.11.0
tzdata==2024.1
urllib3==1.26.18
w3lib==2.1.2
websockets==10.4
zipp==3.18.1
